+++ source/st_lib.c
@@ -66,7 +66,7 @@ void STlib_initNum
   int x,
   int y,
   const patch_t **pl,
-  short* num,
+  int* num,
   boolean* on,
   int     width )
 {
@@ -177,7 +177,7 @@ void STlib_initPercent
   int x,
   int y,
   const patch_t** pl,
-  short* num,
+  int* num,
   boolean* on,
   const patch_t *percent )
 {
